1.使用@Permission进行权限校验 2.解决操作日志记录的问题

+++ modules/core/src/main/java/cn/swallowff/modules/core/system/controller/RoleController.java
@@ -111,6 +111,9 @@ public class RoleController extends BaseController {
     @ResponseBody
     public BaseResp delete(String id){
         BaseResp baseResp = BaseResp.newSuccess();
+        if ("1".equals(id)){
+            return baseResp.putError("无法删除超级管理员角色");
+        }
         int r = roleService.delete(id);
         if (r == 1){
             return baseResp;
@@ -123,6 +126,9 @@ public class RoleController extends BaseController {
         BaseResp baseResp = BaseResp.newSuccess();
         int count = 0;
         for (String id : ids){
+            if ("1".equals(id)){
+                continue;
+            }
             int r = roleService.delete(id);
             if (r == 1){
                 count ++;
